Restructured controllers to begin adding slack bot commands
