+++ app/controllers/slack/commands_controller.rb
@@ -1,0 +1,4 @@
+module Slack
+  class CommandsController < ApplicationController
+  end
+end
