+++ app/models/slack_workspace.rb
@@ -1,3 +1,4 @@
 class SlackWorkspace < ApplicationRecord
   belongs_to :course
+  has_many :slack_users, dependent: :destroy
 end
