+++ db/migrate/20200327234145_add_slack_workspace_id_to_slack_users.rb
@@ -1,0 +1,5 @@
+class AddSlackWorkspaceIdToSlackUsers < ActiveRecord::Migration[5.1]
+  def change
+    add_reference :slack_users, :slack_workspace, foreign_key: true
+  end
+end
